requirements.txt
boto3==1.14.3
botocore==1.17.3
datefinder==0.7.1
numpy==1.21.5
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==9.2.0
pytesseract==0.3.9
pytest==7.1.2
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.36
tweepy==4.10.0
pymysql==1.0.2
beautifulsoup4==4.10.0
selenium==4.4.3
